Fix Pytest plugin empty search paths corner case
Currently when parse_attribute_injection() from the pyszn injection
submodule is called it expects a list of the search paths or None. There
is a corner case where the search paths list is empty, but not None.
This occurs when config.args only contains a path to a file, and not a
directory, causing the isdir() call on "realpath(arg) for arg in
config.args if isdir(arg)" to return False, and then making search_paths
an empty list.

This causes pyszn to present unexpected results when running the
following code:

# Define search paths
if search_paths is None:
    search_paths = [abspath(getcwd())]

This simple fix just assigns None to the search_paths variable when it
is an empty list, matching what pyszn expects.